style: separate packages of data layer
