+++ app/src/main/java/woowacourse/shopping/data/database/cart/CartDatabase.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.database.cart
 
 import android.content.Context
 import androidx.room.Database
