+++ app/src/main/java/woowacourse/shopping/data/database/recent/RecentProductDatabase.kt
@@ -1,4 +1,4 @@
-package woowacourse.shopping.data
+package woowacourse.shopping.data.database.recent
 
 import android.content.Context
 import androidx.room.Database
